Update solution to task 03
Use a helper class Parser for writing the class method parse in the
Directory class. Use attr_reader for adding files and subdirectories
instead of writing separate methods. Use separate hashes for files and subdirectories in the Directory class. Use a helper private function serialize_list for
implementing the serialize method in the Directory class.